# src/main/resources/swagger/employeeAppraisals-api.yml
...
swagger: '2.0'
...
info:
  title: |
    GFT Employee Appraisals API
  description: |
    GFT Employee Appraisals API specification
  version: 1.0.0

# URL Information
host: example.gft.com
schemes:
  - https

# Security Definition
securityDefinitions:
  OAuth2:
    type: oauth2
    flow: accessCode
    authorizationUrl: http://localhost:8888/uaa/oauth/authorize
    tokenUrl: http:/localhost:8888/uaa/oauth/token
    scopes:
      employee: Employee access
      appraisal: Appraisal access
security:
  - OAuth2: [employee, appraisal]

# Endpoints
paths:
...
  '/employees':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Employee
      summary: |
        Returns a list of all employees.
      description: |
        Returns an `Employee` list only if the user has the **Admin** application role.
      operationId: employeesGet
      produces:
        - application/json
      parameters:
        - in: query
          name: searchTerm
          type: string
          required: false
          description: |
            Optional string search term.
        - in: query
          name: page
          type: integer
          required: false
          default: 0
          description: |
            Page number. Defaults to 0.
        - in: query
          name: size
          type: integer
          required: false
          description: |
            Number of elements in each page. Defaults to 10.
          default: 10
      responses:
        '200':
          description: |
            Successful.
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.

    ###########
    # POST ####
    ###########

    post:
      tags:
      - Employee
      summary: |
        Registers a new employee.
      description: |
        Registers a new `Employee`. Users with the **Admin** applicationRole, can register any users. Otherwise they can
        only register themselves and will get an error if they try to register any other users.
      operationId: employeesPost
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: employee
          schema:
            $ref: '#/definitions/EmployeeDTO'
          required: true
          description: |
            `Employee` GFT Login.
      responses:
        '201':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/OperationResultDTO'
          headers:
            Location:
              type: string
              description: |
                URL of the new employee.
        '400':
          description: |
            Input was malformed or did not comply with business rules.
          schema:
            $ref: '#/definitions/OperationResultDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '422':
          description: |
            Entity already exists.
          schema:
            $ref: '#/definitions/OperationResultDTO'

    ###########
    # PUT #####
    ###########

    put:
      tags:
      - Employee
      summary: |
        Registers or updates an employee.
      description: |
        Registers or updates an `Employee`. If the user has the **Admin** applicationRole, they can create or update any
        users. Otherwise they can only create or update themselves and will get an error if they try to create or update
        any other users.
      operationId: employeesPut
      consumes:
        - application/json
      parameters:
        - in: body
          name: employee
          schema:
            $ref: '#/definitions/EmployeeDTO'
          required: true
          description: |
            `Employee` to create or update.
      responses:
        '200':
          description: |
            Successful.
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.

########################################################################################################################
# /employees/{employeeId} ##############################################################################################
########################################################################################################################
  '/employees/{employeeId}':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Employee
      summary: |
        Returns an employee.
      description: |
        Returns an `Employee` with the specified Id. If the user has the **Admin** applicationRole, they can retrieve
        information from any `Employee`. Otherwise they can only retrieve information on themselves.
      operationId: employeesIdGet
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          type: integer
          required: true
          description: |
            `Employee` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/EmployeeDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            An `Employee` with the specified Id was not found.

    ###########
    # PUT #####
    ###########

    put:
      tags:
      - Employee
      summary: |
        Updates a specific employee.
      description: |
        Updates an `Employee`. If the user has the **Admin** applicationRole, they can update any users. Otherwise they
        can only update themselves and will get an error if they try to update any other users.
      operationId: employeesIdPut
      consumes:
        - application/json
      parameters:
        - in: body
          name: employee
          schema:
            $ref: '#/definitions/EmployeeDTO'
          required: true
          description: |
            `Employee` to update.
      responses:
        '200':
          description: |
            Successful.
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            An `Employee` with the specified Id was not found.

    ###########
    # DELETE ##
    ###########

    delete:
      tags:
      - Employee
      summary: |
        Deletes a specific employee.
      description: |
        Deletes an `Employee`. If the user has the **Admin** applicationRole, they can delete any users. Other users
        can't delete themselves.
      operationId: employeesIdDelete
      responses:
        '200':
          description: |
            Successful.
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            An `Employee` with the specified Id was not found.

########################################################################################################################
# /employees/{employeeId}/mentees ######################################################################################
########################################################################################################################
  '/employees/{employeeId}/mentees':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Employee
      summary: |
        Returns the employee mentees.
      description: |
        Returns the mentees associated to an specified `Employee`. An user with the **Admin** application role can see
        the mentees of any `Employee`. An user with the **Mentor** applicationRole can only see their own mentees. Any
        other user can't access this endpoint.
      operationId: employeesIdMenteesGet
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          type: integer
          required: true
          description: |
            `Employee` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            An `Employee` with the specified Id was not found.

########################################################################################################################
# /employees/{employeeId}/mentor #######################################################################################
########################################################################################################################
  '/employees/{employeeId}/mentor':

    ###########
    # GET #####
    ###########

    get:
      tags:
      - Employee
      summary: |
        Returns the mentor associated to the employee.
      description: |
        Returns the mentor associated to an specified `Employee`. An user with the **Admin** application role can see
        the mentor of any `Employee`. Any other user can only obtain his own mentor.
      operationId: employeesIdMentorGet
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          type: integer
          required: true
          description: |
            `Employee` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/EmployeeDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            An `Employee` with the specified Id was not found.

    ###########
    # PUT #####
    ###########

    put:
      tags:
      - Employee
      summary: |
        Associates a mentor to the employee.
      description: |
        Associates a mentor to an specified `Employee`. Can only be used by an user with the **Admin** applicationRole.
      operationId: employeesIdMentorPut
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          type: integer
          required: true
          description: |
            `Employee` Id.
        - in: body
          name: mentor
          schema:
            $ref: '#/definitions/EmployeeDTO'
          required: true
          description: |
            `Employee` Mentor.
      responses:
        '200':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/OperationResultDTO'
        '400':
          description: |
            Input was malformed.
          schema:
            $ref: '#/definitions/OperationResultDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            An `Employee` with the specified Id was not found.

   ###########
    # DELETE #####
    ###########

    delete:
      tags:
      - Employee
      summary: |
        Disassociates a mentor to the employee.
      description: |
        Disassociates a mentor to an specified `Employee`. Can only be used by an user with the **Admin** applicationRole.
      operationId: employeesIdMentorDelete
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          type: integer
          required: true
          description: |
            `Employee` Id.
      responses:
        '200':
          description: |
            Successful.
          schema:
            $ref: '#/definitions/OperationResultDTO'
        '400':
          description: |
            Input was malformed.
          schema:
            $ref: '#/definitions/OperationResultDTO'
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
          schema:
            $ref: '#/definitions/OperationResultDTO'
        '404':
          description: |
            An `Employee` with the specified Id was not found.
          schema:
            $ref: '#/definitions/OperationResultDTO'
        '422':
          description: |
            User does not have a mentor.

########################################################################################################################
# /employees/{employeeId}/appraisals ###################################################################################
########################################################################################################################
...
  '/employees/{employeeId}/relationships/{relationshipId}':

    ###########
    # DELETE ##
    ###########

    delete:
      tags:
      - Employee
      summary: |
        Deletes a Relationship.
      description: |
        Deletes a `Relationship`. Users with the **Mentor** application role, can delete relationships of their mentees.
        Users with the **Admin** application role, can delete any relationships. Users with any other application role
        will receive an error.
      operationId: employeesIdRelationshipsIdDelete
      produces:
        - application/json
      parameters:
        - in: path
          name: employeeId
          required: true
          type: integer
          description: |
            `Employee` Id.
        - in: path
          name: relationshipId
          required: true
          type: integer
          description: |
            `Relationship` Id.
      responses:
        '200':
          description: |
            Successful.
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            An entity with the specified Id wasn't found.

########################################################################################################################
# /relationships/:id ###################################################################################################
########################################################################################################################
  '/relationships':
    get:
      tags:
        - Employee
      summary: |
        Obtains all Relationship entities.
      description: |
        Fetches all Relationship entities from database. Refer to the database model for details.
      operationId: relationshipsGet
      produces:
        - application/json
      responses:
        '200':
          schema:
            type: array
            items:
              $ref: '#/definitions/RelationshipTypeDTO'
          description: |
            Successful.
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.

########################################################################################################################
# /relationships/:id ###################################################################################################
########################################################################################################################
  '/relationships/{relationshipId}':
    get:
      tags:
        - Employee
      summary: |
        Obtains a Relationship entity.
      description: |
        Fetches a single Relationship entity from database. Refer to the database model for details.
      operationId: relationshipsIdGet
      produces:
        - application/json
      parameters:
        - in: path
          name: relationshipId
          required: true
          type: integer
          description: |
            `Relationship` Id.
      responses:
        '200':
          schema:
            $ref: '#/definitions/RelationshipTypeDTO'
          description: |
            Successful.
        '403':
          description: |
            The user isn't authenticated or authorized to perform this action.
        '404':
          description: |
            An entity with the specified Id wasn't found.

########################################################################################################################
# Definitions ##########################################################################################################
########################################################################################################################

#
# Note: Keep the definitions ordered alphabetically!
#
definitions:
  ApplicationRoleDTO:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
    required:
      - id
...
  EmployeeDTO:
    type: object
    properties:
      id:
        type: integer
      jobLevel:
        $ref: '#/definitions/JobLevelDTO'
      applicationRole:
        $ref: '#/definitions/ApplicationRoleDTO'
      email:
        type: string
      gftIdentifier:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      isAdmin:
        type: boolean
      isMentor:
        type: boolean
      isPeer:
        type: boolean
...
  FieldErrorDTO:
    type: object
    properties:
      field:
        type: string
      message:
        type: string
...
  JobLevelDTO:
    type: object
    properties:
      id:
        type: integer
      jobFamily:
        $ref: '#/definitions/JobFamilyDTO'
      name:
        type: string
      description:
        type: string
      expertise:
        type: string
    required:
      - id
...
  OperationResultDTO:
    type: object
    properties:
      message:
        type: string
      data:
        type: object
      errors:
        type: array
        items:
          $ref: '#/definitions/FieldErrorDTO'
...
  RelationshipTypeDTO:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
...